+++ src/tests/test_imdb.py
@@ -312,101 +312,45 @@ def test_get_title_episodes_raises_imdb_id_is_not_that_of_a_tv_show(client):
         client.get_title_episodes(non_show_imdb_id)
 
 
-@pytest.fixture(scope='module')
-def client():
-    client = Imdb(locale='en_US')
-    yield client
-    client.clear_cached_credentials()
-
-
-@pytest.mark.parametrize('p', [
-    {
+@pytest.mark.parametrize(
+    'params, exp_num_episodes, exp_episodes_total, exp_seasons',
+    [
         # firefly
-        'imdb_id': 'tt0303461',
-        'start': 0,
-        'end': 500,
-        'season': 0,
-        'region': None,
-        'expected_episodes_length': 14,
-        'expected_episodes_total': 14,
-        'expected_all_seasons': [1],
-        'expected_keys': [
-            'allSeasons', 'end', 'episodes', 'region', 'season',
-            'seriesTitle', 'start', 'totalEpisodes'
-        ]
-    },
-    {
-        # Naruto shippuuden
-        'imdb_id': 'tt0988824',
-        'start': 0,
-        'end': 50,
-        'season': 0,
-        'region': None,
-        'expected_episodes_length': 50,
-        'expected_episodes_total': 500,
-        'expected_all_seasons': [1, None],
-        'expected_keys': [
-            'allSeasons', 'end', 'episodes', 'region', 'season',
-            'seriesTitle', 'start', 'totalEpisodes'
-        ]
-    },
-    {
-        # Naruto shippuuden
-        'imdb_id': 'tt0988824',
-        'start': 120,
-        'end': 140,
-        'season': 0,
-        'region': None,
-        'expected_episodes_length': 20,
-        'expected_episodes_total': 500,
-        'expected_all_seasons': [1, None],
-        'expected_keys': [
-            'allSeasons', 'end', 'episodes', 'region', 'season',
-            'seriesTitle', 'start', 'totalEpisodes'
-        ]
-    },
-    {
+        (
+            dict(imdb_id='tt0303461', offset=0, limit=500, season=1),
+            14,
+            14,
+            [1],
+        ),
         # Breaking bad
-        'imdb_id': 'tt0903747',
-        'start': 0,
-        'end': 6,
-        'season': 3,
-        'region': None,
-        'expected_episodes_length': 6,
-        'expected_episodes_total': 13,
-        'expected_all_seasons': [1, 2, 3, 4, 5],
-        'expected_keys': [
-            'allSeasons', 'end', 'episodes', 'region', 'season',
-            'seriesTitle', 'start', 'totalEpisodes'
-        ]
-    },
-    {
+        (
+            dict(imdb_id='tt0903747', offset=1, limit=6, season=4),
+            5,
+            13,
+            [1, 2, 3, 4, 5],
+        ),
         # Detective conan
-        'imdb_id': 'tt0131179',
-        'start': 0,
-        'end': 500,
-        'season': 45,
-        'region': None,
-        'expected_episodes_length': 4,
-        'expected_episodes_total': 4,
-        'expected_all_seasons': list(range(1, 47)),
-        'expected_keys': [
-            'allSeasons', 'end', 'episodes', 'region', 'season',
-            'seriesTitle', 'start', 'totalEpisodes'
-        ]
-    }
-])
-def test_get_title_episodes_detailed(p, client):
+        (
+            dict(imdb_id='tt0131179', offset=0, limit=500, season=46),
+            4,
+            4,
+            list(range(1, 47)),
+        ),
+    ]
+)
+def test_get_title_episodes_detailed(
+    params, client, exp_num_episodes, exp_episodes_total, exp_seasons
+):
+    expected_keys = [
+        'allSeasons', 'end', 'episodes', 'region', 'season', 'seriesTitle',
+        'start', 'totalEpisodes'
+    ]
+    resource = client.get_title_episodes_detailed(**params)
 
-    resource = client.get_title_episodes_detailed(
-        p['imdb_id'], p['end'], region=p['region'],
-        season=p['season'], start=p['start']
-    )
-
-    assert sorted(resource.keys()) == sorted(p['expected_keys'])
-    assert len(resource['episodes']) == p['expected_episodes_length']
-    assert resource['totalEpisodes'] == p['expected_episodes_total']
-    assert resource['allSeasons'] == p['expected_all_seasons']
+    assert sorted(resource.keys()) == sorted(expected_keys)
+    assert len(resource['episodes']) == exp_num_episodes
+    assert resource['totalEpisodes'] == exp_episodes_total
+    assert resource['allSeasons'] == exp_seasons
 
 
 def test_get_name_images(client):